Add vmIntrinsics.cpp. Fix commit author.
